Int, sorry. Float exist in function.

+++ src/pocketmine/level/generator/biome/BiomeSelector.php
@@ -113,8 +113,8 @@ class BiomeSelector{
 	 * @return Biome
 	 */
 	public function pickBiome($x, $z){
-		$temperature = (float) ($this->getTemperature($x, $z) * 63);
-		$rainfall = (float) ($this->getRainfall($x, $z) * 63);
+		$temperature = (int) ($this->getTemperature($x, $z) * 63);
+		$rainfall = (int) ($this->getRainfall($x, $z) * 63);
 
 		$biomeId = $this->map[$temperature + ($rainfall << 6)];
 		return isset($this->biomes[$biomeId]) ? $this->biomes[$biomeId] : $this->fallback;
